requirements.txt
flask==3.1.0
biopython==1.85
requests==2.32.3
